chore: remove priority and refactor code
